Create `SinglePlotFarm::new()` and move instantiation there
